requirements.txt
PyYAML
selenium
robotframework
